+++ frontend-hybrid/yakal/lib/screens/Calender/calender_screen.dart
@@ -1,14 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:yakal/widgets/Calender/pill_custom_calendar.dart';
 import 'package:yakal/widgets/Home/pill_todo_layout.dart';
 
 import '../../viewModels/Calendar/calendar_viewmodel.dart';
 import '../../widgets/Calender/calendar_info_layout.dart';
-import '../../widgets/Calender/pill_calender.dart';
+import '../../widgets/Calender/pill_calendar.dart';
 
 class CalenderScreen extends StatelessWidget {
-  final CalendarViewModel viewModel = Get.put(CalendarViewModel());
   CalenderScreen({Key? key}) : super(key: key);
+
+  final CalendarViewModel viewModel = Get.put(CalendarViewModel());
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +33,8 @@ class CalenderScreen extends StatelessWidget {
       ),
       body: Column(
         children: [
-          PillCalender(viewModel),
+          // CustomPillCalendar(viewModel: viewModel),
+          PillCalendar(viewModel: viewModel),
           Container(
             width: double.infinity,
             height: 2,
